ciborium::de

Function from_reader_with_recursion_limit

pub fn from_reader_with_recursion_limit<T: DeserializeOwned, R: Read>(
    reader: R,
    recurse_limit: usize,
) -> Result<T, Error<R::Error>>
where
    R::Error: Debug,

Deserializes as CBOR from a type with impl ciborium_io::Read, with a specified maximum recursion limit. Inputs that are nested beyond the specified limit will result in Error::RecursionLimitExceeded .

Set a high recursion limit at your own risk (of stack exhaustion)!